requirement.txt
pymysql==1.0.2
cryptography==3.4.7
Django==3.1.9
django-simpleui==2021.6.2
Pillow==8.2.0
djangorestframework==3.12.4
pygments==2.9.0
httpie==2.4.0
